+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.0.beta3 2015-??-??
+ - 2015-08-17 Added AgentTicketBulk screen capability load plugin modules.
  - 2015-08-13 Added switchable display of sender and recipients (realname or full field incl. email address) in ticket zoom for agent and customer interface.
 
 #5.0.0.beta2 2015-08-18